Refactor: Improve order creation and entity robustness
This commit introduces several improvements to the order handling logic and related entities:

1.  **Refactored `CreateOrderUseCase`:**
    *   I removed a hardcoded URL for fetching offer data, now correctly using `offer_repository.find_offer`.
    *   I ensured an `is_active()` check is performed on offers before applying them.
    *   I removed an unnecessary random sleep that was present when an offer was applied.

2.  **Improved Error Handling in `Order` Entity:**
    *   I replaced generic `Exception` throws in the `Order` entity's validation logic with more specific exceptions (`ValueError`, `TypeError`).

3.  **Standardized Monetary Values to `Decimal`:**
    *   I changed `total_price` in the `Order` entity to use `Decimal` for better precision.
    *   I changed `total_price` in the `Cart` entity to `Decimal`.
    *   I changed `discount_value` in the `Offer` entity to `Decimal`.
    *   I updated `apply_discount` in `Offer` to work with `Decimal`.
    *   I modified `CreateOrderOutputDto` to expect `Decimal` for `total_price`.
    *   SQLAlchemy models (`OfferModel`, `CartModel`) were updated to use `Numeric(10, 2)` for these fields, ensuring database compatibility. `OrderModel` already used `Numeric`.

4.  **Reviewed Payment Creation Logic:**
    *   I confirmed that `payment_method` and `payment_card_gateway` in `CreateOrderUseCase` remain `None` upon initial payment creation, as there's no DTO input to determine them at this stage. A recommendation for future DTO enhancement was noted.

5.  **Updated Unit Tests:**
    *   All relevant unit tests for `Order`, `Cart`, `Offer` entities, and the `CreateOrderUseCase` were updated to reflect the changes to `Decimal` usage and new exception types. This included fixing `src/domain/order/test_order_entity.py`.

These changes enhance the reliability, maintainability, and accuracy of the order processing components.

+++ src/infrastructure/cart/sqlalchemy/cart_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import (Column, Float, ForeignKey)
+from sqlalchemy import (Column, Float, ForeignKey, Numeric)
 from sqlalchemy.dialects.postgresql import UUID
 
 from infrastructure.api.database import Base
@@ -9,4 +9,4 @@ class CartModel(Base):
 
     id = Column(UUID, primary_key=True, index=True)
     user_id = Column(UUID, ForeignKey("tb_users.id"))
-    total_price = Column(Float, nullable=False)
+    total_price = Column(Numeric(10, 2), nullable=False)  # Changed from Float
